requirements.txt
av~=8.1.0
pillow~=8.3.2
